Split match stuff into its own file
